Remove unused gitbutler-repo-actions crate
