+++ Cargo.toml
@@ -193,7 +193,6 @@ gitbutler-filemonitor = { path = "crates/gitbutler-filemonitor" }
 gitbutler-branch-actions = { path = "crates/gitbutler-branch-actions" }
 gitbutler-oplog = { path = "crates/gitbutler-oplog" }
 gitbutler-repo = { path = "crates/gitbutler-repo" }
-gitbutler-repo-actions = { path = "crates/gitbutler-repo-actions" }
 gitbutler-project = { path = "crates/gitbutler-project" }
 gitbutler-user = { path = "crates/gitbutler-user" }
 gitbutler-branch = { path = "crates/gitbutler-branch" }
