+++ crates/gitbutler-branch-actions/Cargo.toml
@@ -30,7 +30,6 @@ but-meta = { workspace = true, features = ["legacy"] }
 
 gitbutler-oplog.workspace = true
 gitbutler-repo.workspace = true
-gitbutler-repo-actions.workspace = true
 gitbutler-branch.workspace = true
 gitbutler-reference.workspace = true
 gitbutler-commit.workspace = true
