+++ crates/gitbutler-branch-actions/src/branch_manager/branch_creation.rs
@@ -1,8 +1,7 @@
 use anyhow::{Context as _, Result};
-use but_ctx::access::RepoExclusive;
+use but_ctx::{Context, access::RepoExclusive};
 use gitbutler_branch::{self, BranchCreateRequest, dedup};
 use gitbutler_oplog::SnapshotExt;
-use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::Stack;
 use tracing::instrument;
 
@@ -49,10 +48,34 @@ impl BranchManager<'_> {
         let branch = Stack::new_empty(self.ctx, name, target_base_oid, order)?;
 
         vb_state.set_stack(branch.clone())?;
-        self.ctx.add_branch_reference(&branch)?;
+        ensure_stack_reference(self.ctx, &branch)?;
 
         crate::integration::update_workspace_commit_with_vb_state(&vb_state, self.ctx, false)?;
 
         Ok(branch)
     }
+}
+
+fn ensure_stack_reference(ctx: &Context, stack: &Stack) -> Result<()> {
+    let repo = ctx.repo.get()?;
+    let refname = stack.refname()?.to_string();
+    let head_oid = stack.head_oid(ctx)?;
+    let previous = match repo
+        .try_find_reference(&refname)
+        .context("failed to lookup reference")?
+    {
+        Some(reference) => {
+            if reference.id() == head_oid {
+                return Ok(());
+            }
+            gix::refs::transaction::PreviousValue::Any
+        }
+        None => gix::refs::transaction::PreviousValue::MustNotExist,
+    };
+
+    let refname: gix::refs::FullName = refname.as_str().try_into()?;
+    repo.reference(refname, head_oid, previous, "new vbranch")
+        .context("failed to create branch reference")?;
+
+    Ok(())
 }
