--- crates/gitbutler-repo-actions/src/lib.rs
@@ -1,2 +1,0 @@
-mod repository;
-pub use repository::RepoActionsExt;
